+++ index.ts
@@ -7,7 +7,6 @@ import xDownloadOptions from "./middlewares/x-download-options";
 import xFrameOptions from "./middlewares/x-frame-options";
 import xPermittedCrossDomainPolicies from "./middlewares/x-permitted-cross-domain-policies";
 import xPoweredBy from "./middlewares/x-powered-by";
-import depd = require("depd");
 
 interface HelmetOptions {
   contentSecurityPolicy?: any;
@@ -21,14 +20,11 @@ interface HelmetOptions {
   permittedCrossDomainPolicies?: any;
   referrerPolicy?: any;
   xssFilter?: any;
-  noCache?: any;
 }
 
 interface MiddlewareFunction {
   (req: IncomingMessage, res: ServerResponse, next: () => void): void;
 }
-
-const deprecate = depd("helmet");
 
 const DEFAULT_MIDDLEWARE = [
   "dnsPrefetchControl",
@@ -51,8 +47,7 @@ type MiddlewareName =
   | "noSniff"
   | "permittedCrossDomainPolicies"
   | "referrerPolicy"
-  | "xssFilter"
-  | "noCache";
+  | "xssFilter";
 
 const middlewares: MiddlewareName[] = [
   "contentSecurityPolicy",
@@ -66,7 +61,6 @@ const middlewares: MiddlewareName[] = [
   "permittedCrossDomainPolicies",
   "referrerPolicy",
   "xssFilter",
-  "noCache",
 ];
 
 function helmet(options: Readonly<HelmetOptions> = {}) {
@@ -137,10 +131,5 @@ helmet.noSniff = xContentTypeOptions;
 helmet.permittedCrossDomainPolicies = xPermittedCrossDomainPolicies;
 helmet.referrerPolicy = referrerPolicy;
 helmet.xssFilter = require("x-xss-protection");
-
-helmet.noCache = deprecate.function(
-  require("nocache"),
-  "helmet.noCache is deprecated and will be removed in helmet@4. You can use the `nocache` module instead. For more, see https://github.com/helmetjs/helmet/issues/215."
-);
 
 export = helmet;