+++ test/index.test.ts
@@ -1,8 +1,5 @@
 import helmet = require("..");
 
-import { IncomingMessage, ServerResponse } from "http";
-import connect = require("connect");
-import request = require("supertest");
 import expectCt from "../middlewares/expect-ct";
 import referrerPolicy from "../middlewares/referrer-policy";
 import xContentTypeOptions from "../middlewares/x-content-type-options";
@@ -54,39 +51,6 @@ describe("helmet", function () {
       expect(helmet.ieNoOpen.name).toBe(xDowloadOptions.name);
     });
 
-    // This test will be removed in helmet@4.
-    it("calls through to nocache but emits a deprecation warning", function () {
-      const deprecationPromise = new Promise((resolve) => {
-        process.once("deprecation", (deprecationError) => {
-          expect(
-            deprecationError.message.indexOf(
-              "You can use the `nocache` module instead."
-            ) !== -1
-          ).toBeTruthy();
-          resolve();
-        });
-      });
-
-      const app = connect();
-      app.use(helmet.noCache());
-      app.use((_req: IncomingMessage, res: ServerResponse) => {
-        res.end("Hello world!");
-      });
-      const supertestPromise = request(app)
-        .get("/")
-        .expect(200)
-        .expect("Surrogate-Control", "no-store")
-        .expect(
-          "Cache-Control",
-          "no-store, no-cache, must-revalidate, proxy-revalidate"
-        )
-        .expect("Pragma", "no-cache")
-        .expect("Expires", "0")
-        .expect("Hello world!");
-
-      return Promise.all([deprecationPromise, supertestPromise]);
-    });
-
     it("aliases the Referrer-Policy middleware to helmet.referrerPolicy", () => {
       expect(helmet.referrerPolicy.name).toBe(referrerPolicy.name);
     });
@@ -107,7 +71,6 @@ describe("helmet", function () {
       jest.spyOn(helmet, "hsts");
       jest.spyOn(helmet, "hsts");
       jest.spyOn(helmet, "ieNoOpen");
-      jest.spyOn(helmet, "noCache");
       jest.spyOn(helmet, "noSniff");
       jest.spyOn(helmet, "permittedCrossDomainPolicies");
       jest.spyOn(helmet, "referrerPolicy");
@@ -135,7 +98,6 @@ describe("helmet", function () {
 
       expect(helmet.contentSecurityPolicy).not.toHaveBeenCalled();
       expect(helmet.expectCt).not.toHaveBeenCalled();
-      expect(helmet.noCache).not.toHaveBeenCalled();
       expect(helmet.permittedCrossDomainPolicies).not.toHaveBeenCalled();
     });
 
@@ -158,7 +120,6 @@ describe("helmet", function () {
       expect(helmet.xssFilter).toHaveBeenCalledWith({});
       expect(helmet.contentSecurityPolicy).not.toHaveBeenCalled();
       expect(helmet.expectCt).not.toHaveBeenCalled();
-      expect(helmet.noCache).not.toHaveBeenCalled();
     });
 
     it("lets you enable a normally-disabled middleware", function () {
@@ -183,7 +144,6 @@ describe("helmet", function () {
       expect(helmet.xssFilter).toHaveBeenCalledWith({});
       expect(helmet.contentSecurityPolicy).not.toHaveBeenCalled();
       expect(helmet.expectCt).not.toHaveBeenCalled();
-      expect(helmet.noCache).not.toHaveBeenCalled();
     });
 
     it("lets you set options for a default middleware", function () {
@@ -208,7 +168,6 @@ describe("helmet", function () {
       expect(helmet.xssFilter).toHaveBeenCalledWith({});
       expect(helmet.contentSecurityPolicy).not.toHaveBeenCalled();
       expect(helmet.expectCt).not.toHaveBeenCalled();
-      expect(helmet.noCache).not.toHaveBeenCalled();
       expect(helmet.permittedCrossDomainPolicies).not.toHaveBeenCalled();
     });
 
@@ -239,7 +198,6 @@ describe("helmet", function () {
       expect(helmet.noSniff).toHaveBeenCalledWith({});
       expect(helmet.xssFilter).toHaveBeenCalledWith({});
       expect(helmet.expectCt).not.toHaveBeenCalled();
-      expect(helmet.noCache).not.toHaveBeenCalled();
       expect(helmet.permittedCrossDomainPolicies).not.toHaveBeenCalled();
     });