+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Removed
 
-- `helmet.hpkp`. If you still need it, use the `hpkp` package on npm.
 - `helmet.featurePolicy`. If you still need it, use the `feature-policy` package on npm.
+- `helmet.hpkp`. If you still need it, use the `hpkp` package on npm.
+- `helmet.noCache`. If you still need it, use the `nocache` package on npm.
 
 ## Unreleased